修改频道内容adapter
